# Speedup take_bytes (-35% -69%) by precalculating capacity