--- a/arrow-select/src/take.rs
+++ b/arrow-select/src/take.rs
@@ -516,14 +516,27 @@ fn take_bytes<T: ByteArrayType, IndexType: ArrowPrimitiveType>(
         };
         return Ok(array);
     } else if array.null_count() == 0 {
-        offsets.extend(indices.values().iter().enumerate().map(|(i, index)| {
+        let mut capacity = 0;
+        for index in indices.values() {
+            let index = index.as_usize();
+            if indices.is_valid(index) {
+                capacity += input_offsets[index + 1].as_usize() - input_offsets[index].as_usize();
+                offsets.push(T::Offset::from_usize(capacity).expect("overflow"));
+            }
+        }
+        values = MutableBuffer::new(capacity);
+
+        for (i, index) in indices.values().iter().enumerate() {
             if indices.is_valid(i) {
                 let s: &[u8] = array.value(index.as_usize()).as_ref();
                 values.extend_from_slice(s);
             }
-            T::Offset::usize_as(values.len())
-        }));
-        nulls = indices.nulls().map(|b| b.inner().sliced());
+        };
+        let array = unsafe {
+            let offsets = OffsetBuffer::new_unchecked(offsets.into());
+            GenericByteArray::<T>::new_unchecked(offsets, values.into(), None)
+        };
+        return Ok(array);
     } else {
         let num_bytes = bit_util::ceil(data_len, 8);