# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->


Closes #7432

# Rationale for this change
 
Performance improvements:

<details>


```
Benchmarking take str 512: Collecting 100 samples in estimated 5.0014 s (1.3M ittake str 512            time:   [3.9906 µs 3.9937 µs 3.9970 µs]
                        change: [-45.040% -44.844% -44.657%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

Benchmarking take str 1024: Collecting 100 samples in estimated 5.0285 s (556k itake str 1024           time:   [9.0347 µs 9.0435 µs 9.0519 µs]
                        change: [-69.387% -69.048% -68.716%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Benchmarking take str null indices 512: Collecting 100 samples in estimated 5.00take str null indices 512
                        time:   [2.4167 µs 2.4209 µs 2.4269 µs]
                        change: [-48.977% -48.536% -48.234%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  3 (3.00%) high mild
  6 (6.00%) high severe

Benchmarking take str null indices 1024: Collecting 100 samples in estimated 5.0take str null indices 1024
                        time:   [5.9829 µs 5.9940 µs 6.0151 µs]
                        change: [-57.375% -54.878% -52.153%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) low severe
  2 (2.00%) high mild
  3 (3.00%) high severe

Benchmarking take str null values 1024: Collecting 100 samples in estimated 5.00take str null values 1024
                        time:   [6.1283 µs 6.1357 µs 6.1418 µs]
                        change: [-23.035% -22.635% -22.315%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  10 (10.00%) low severe
  2 (2.00%) low mild
  1 (1.00%) high mild
  9 (9.00%) high severe

Benchmarking take str null values null indices 1024: Collecting 100 samples in etake str null values null indices 1024
                        time:   [5.2781 µs 5.2905 µs 5.3024 µs]
                        change: [-9.2518% -8.9305% -8.6089%] (p = 0.00 < 0.05)
                        Performance has improved.

```


</details>
# What changes are included in this PR?


<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please call them out.
-->