@Dandandan 2025-04-25
> Do you see the same performance regression for `take stringview`? I'm still just digging into the code to see if that is even possible.
> 
> Main:
> 
> ```
> take stringview 512     time:   [287.26 ns 287.75 ns 288.28 ns]
> 
> take stringview 1024    time:   [489.07 ns 489.99 ns 491.18 ns]
> ```
> 
> PR:
> 
> ```
> take stringview 512     time:   [360.36 ns 361.02 ns 361.95 ns]
> 
> take stringview 1024    time:   [633.66 ns 633.79 ns 633.98 ns]
> ```

I am not at a machine right now to test, but StringView doesn't use this code path.