@mbutrovich 2025-04-25
> I am not at a machine right now to test, but StringView doesn't use this code path.

Generated code looks the same, so must be some sort of noise on my machine.