@alamb 2025-04-29
> > I think at this point there is little point in using `MutableBuffer` over `Vec` as the latter provides more performant (specialization over `T`, better inlining), slightly more safe and more complete API. The same probably applies for a lot of `Builder`-type APIs probably.
> > It might be worth spending some time documenting the best and most performant way to construct / transform Arrow arrays and apply it accross the arrow-rs create (marking stuff as deprecated if needed).
> 
> This LGTM and is a great improvement. I also would love to see this documentation if you don't mind dumping your thoughts on the topic while they're fresh!

- Filed https://github.com/apache/arrow-rs/issues/7455